added logout controller + refactor
