+++ src/tests/user/userService.test.ts
@@ -103,3 +103,4 @@ describe('getUser', () => {
 
       });
 });
+
